Add: 모델 간 FK 참조 관계 설정
migrations/ 각 모델 migration 파일 내 명시

*Command*
npx sequelize db:migrate // migrations 폴더 내 모든 migrate 의 up 커맨드 실행(테이블 생성)
                         // 이미 테이블이 존재한다면 갱신하지 않음
npx sequelize db:migrate:undo // migrate를 되돌림
npx sequelize db:migrate:undo:all 모든 테이블들을 삭제

ex) npx sequelize db:migrate:undo:all
    npx sequelize db:migrate

+++ api-server/migrations/20191113061728-create-user.js
@@ -1,5 +1,3 @@
-
-
 module.exports = {
   up: (queryInterface, Sequelize) => {
     return queryInterface.createTable('Users', {
@@ -7,7 +5,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT(11),
       },
       name: {
         type: Sequelize.STRING,
