+++ api-server/migrations/20191113062914-create-video.js
@@ -5,7 +5,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT(11),
       },
       name: {
         type: Sequelize.STRING,
